Update python-main.js
Made search function search for #l

Changed the search function to search for #l instead of just hash, so if any other hash parameters are added it will ignore them.

Depending on how hash parameters are formatted, this might need to be changed, to make it just search for hashes and then split between &s like queries.