Moved website translation block
Moved the block that changes the language of the page based on a URL query.
It was in the setupButtons function because the code was copied from the click handler of the translate button. The block has been moved to the translations function and is called at the end of the web_editor function.

Also removed the get_language_functino as we have decided to use URL queries instead of URL fragments